chore: fix tests using parent account (#11616)
* chore: fix tests using parent account

* chore: fix return, trim list

* chore: try this

* chore: fix tests

* chore: these are insanely long on windows

* chore: try this

* chore: try this

* chore: try this

* chore: remove race

* chore: remove race

* chore: try this

* chore: try this

* chore: fix this
